+++ posthog-android/src/main/java/com/posthog/android/PostHogAndroid.kt
@@ -13,9 +13,11 @@ import com.posthog.android.internal.PostHogAndroidLogger
 import com.posthog.android.internal.PostHogAndroidNetworkStatus
 import com.posthog.android.internal.PostHogAppInstallIntegration
 import com.posthog.android.internal.PostHogLifecycleObserverIntegration
+import com.posthog.android.internal.PostHogMetaPropertiesApplier
 import com.posthog.android.internal.PostHogSharedPreferences
 import com.posthog.android.internal.appContext
 import com.posthog.android.internal.getPackageInfo
+import com.posthog.android.internal.versionCodeCompat
 import com.posthog.android.replay.PostHogReplayIntegration
 import com.posthog.android.replay.internal.PostHogLogCatIntegration
 import com.posthog.android.surveys.PostHogSurveysIntegration
@@ -78,13 +80,16 @@ public class PostHogAndroid private constructor() {
 
             val packageInfo = getPackageInfo(context, config)
             val packageName = packageInfo?.packageName ?: ""
+            val versionName = packageInfo?.versionName ?: ""
+            val buildNumber = packageInfo?.versionCodeCompat() ?: 0L
 
             // only frames coming from the package name will be considered inApp by default
             if (packageName.isNotEmpty() && !packageName.startsWith("android.")) {
                 config.errorTrackingConfig.inAppIncludes.add(packageName)
             }
 
-            config.context = config.context ?: PostHogAndroidContext(context, config)
+            val androidContext = config.context ?: PostHogAndroidContext(context, config)
+            config.context = androidContext
 
             val legacyPath = context.getDir("app_posthog-disk-queue", Context.MODE_PRIVATE)
             val path = File(context.cacheDir, "posthog-disk-queue")
@@ -107,6 +112,10 @@ public class PostHogAndroid private constructor() {
                 config.sdkName = "posthog-android"
                 config.sdkVersion = BuildConfig.VERSION_NAME
             }
+
+            val releaseIdentifierFallback = "$packageName@$versionName+$buildNumber"
+            val metaPropertiesApplier = PostHogMetaPropertiesApplier()
+            metaPropertiesApplier.applyToConfig(context, config, releaseIdentifierFallback)
 
             val mainHandler = MainHandler()
             config.addIntegration(PostHogReplayIntegration(context, config, mainHandler))