+++ posthog/src/main/java/com/posthog/internal/errortracking/ThrowableCoercer.kt
@@ -25,6 +25,7 @@ public class ThrowableCoercer {
     public fun fromThrowableToPostHogProperties(
         throwable: Throwable,
         inAppIncludes: List<String> = listOf(),
+        releaseIdentifier: String? = null,
     ): MutableMap<String, Any> {
         val exceptions = mutableListOf<Map<String, Any>>()
         val throwableList = mutableListOf<Throwable>()
@@ -70,6 +71,11 @@ public class ThrowableCoercer {
                     myFrame["module"] = frame.className
                     myFrame["function"] = frame.methodName
                     myFrame["platform"] = "java"
+
+                    // add release identifier for symbolication
+                    if (!releaseIdentifier.isNullOrEmpty()) {
+                        myFrame["map_id"] = releaseIdentifier
+                    }
 
                     if (frame.lineNumber >= 0) {
                         myFrame["lineno"] = frame.lineNumber