+++ scripts/bump-version.sh
@@ -25,12 +25,16 @@ case "$MODULE" in
   "server")
     perl -pi -e "s/serverVersion=.*/serverVersion=$NEW_VERSION/g" $GRADLE_FILEPATH
     ;;
+  "androidPlugin")
+    perl -pi -e "s/androidPluginVersion=.*/androidPluginVersion=$NEW_VERSION/g" $GRADLE_FILEPATH
+    ;;
   *)
     echo "Usage: $0 {core|android|server} <version>"
     echo "Examples:"
     echo "  $0 core 3.0.1"
     echo "  $0 android 3.0.2"
     echo "  $0 server 1.0.1"
+    echo "  $0 androidPlugin 3.0.2"
     exit 1
     ;;
 esac