+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Changelogs can be found in the individual packages:
 - [posthog](./posthog/CHANGELOG.md)
 - [posthog-android](./posthog-android/CHANGELOG.md)
 - [posthog-server](./posthog-server/CHANGELOG.md)
+- [posthog-android-gradle-plugin](./posthog-android-gradle-plugin/CHANGELOG.md)